Refactored tree_template.dart.
