+++ lib/routes/explore_routes.dart
@@ -4,7 +4,7 @@ import 'app_routes.dart';
 // Import page files
 import 'package:httapp/pages/explore/explore.dart';
 import 'package:httapp/pages/explore/tree_list.dart';
-import 'package:httapp/ui/tree_template.dart';
+import 'package:httapp/ui/template_tree_page.dart';
 
 class ExploreRoutes implements AppRoutes{
   // Route name constants
@@ -29,7 +29,7 @@ class ExploreRoutes implements AppRoutes{
     for (var page in externalRoutes.entries) {
       // add '/treePage/' before each key as to not confuse what the route is for
       final routeName = '/${page.key}';
-      routes[routeName] = (context) => TreeTemplatePage(
+      routes[routeName] = (context) => TemplateTreePage(
           id: page.value['id'], 
           name: page.value['name'], 
           body: page.value['body'],
